+++ src/pyventim/__init__.py
@@ -3,3 +3,6 @@
 <hr>
 .. include:: EXAMPLES.md
 """
+
+from .utils import utils
+from .exploration import Exploration
