+++ tests/test_exploration.py
@@ -5,8 +5,9 @@ import datetime
 from typing import Dict, List, Any
 
 # Module to test
-from pyventim.public import EventimExploration
+from pyventim import Exploration
 
+Exploration
 # Setup
 SEARCH_TERM = "The"
 CATEGORIES = ["Musical & Show|Musical"]
@@ -24,7 +25,7 @@ NEXT_WEEK = TODAY + datetime.timedelta(days=7)
 ### Query parameter validation tests ###
 ########################################
 def test_validate_required_parameters():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(
         ValueError,
         match="Must have search_term, categories or city_ids in the query parameters!",
@@ -38,7 +39,7 @@ def test_validate_required_parameters():
 
 
 def test_validate_page_parameter():  # pylint: disable=C0116
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(
         ValueError,
         match="page must be a positive integer!",
@@ -51,7 +52,7 @@ def test_validate_page_parameter():  # pylint: disable=C0116
 
 
 def test_validate_sort_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(
         ValueError,
         match="sort must match: DateAsc, DateDesc, NameAsc, NameDesc, Rating, Recommendation",
@@ -67,7 +68,7 @@ def test_validate_sort_parameter():
 
 
 def test_validate_search_term_parameter():  # pylint: disable=C0116
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(
         ValueError,
         match="search_term must have atleast two characters!",
@@ -78,7 +79,7 @@ def test_validate_search_term_parameter():  # pylint: disable=C0116
 
 
 def test_validate_categories_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="categories must be a list of strings!"):
         explorer.validate_categories_parameter("This is not category value.")
         explorer.validate_categories_parameter(["This is valid", False])
@@ -87,7 +88,7 @@ def test_validate_categories_parameter():
 
 
 def test_validate_city_ids_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="city_ids must be a list of integers!"):
         explorer.validate_city_ids_parameter("This is not category value.")
         explorer.validate_city_ids_parameter(["This is not valid valid", 123])
@@ -96,7 +97,7 @@ def test_validate_city_ids_parameter():
 
 
 def test_validate_date_from_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="date_from must be of type datetime.date!"):
         explorer.validate_date_from_parameter("This is not a valid value.")
 
@@ -104,7 +105,7 @@ def test_validate_date_from_parameter():
 
 
 def test_validate_date_to_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="date_to must be of type datetime.date!"):
         explorer.validate_date_to_parameter("This is not a valid value.")
 
@@ -112,7 +113,7 @@ def test_validate_date_to_parameter():
 
 
 def test_validate_time_from_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="time_from must be of type datetime.time!"):
         explorer.validate_time_from_parameter("This is not a valid value.")
 
@@ -120,7 +121,7 @@ def test_validate_time_from_parameter():
 
 
 def test_validate_time_to_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="time_to must be of type datetime.time!"):
         explorer.validate_time_to_parameter("This is not a valid value.")
 
@@ -128,7 +129,7 @@ def test_validate_time_to_parameter():
 
 
 def test_validate_in_stock_parameter():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     with pytest.raises(ValueError, match="in_stock must be of type bool!"):
         explorer.validate_in_stock_parameter("This is not a valid value.")
 
@@ -139,7 +140,7 @@ def test_validate_in_stock_parameter():
 ### Query builder validation tests ###
 ######################################
 def test_build_query_parameters():
-    explorer: EventimExploration = EventimExploration()
+    explorer: Exploration = Exploration()
     inputs = {
         "search_term": SEARCH_TERM,
         "date_from": TODAY,
